*create-select works *flow of create-select apartments
