+++ apps/client/src/routes/select-apartment.tsx
@@ -1,11 +1,11 @@
 import { createFileRoute, redirect } from '@tanstack/react-router';
-import CreateApartment from '../pages/CreateApartment';
+import { SelectApartment } from '../pages/SelectApartment';
 
-export const Route = createFileRoute('/create-apartment')({
+export const Route = createFileRoute('/select-apartment')({
   beforeLoad: ({ context }) => {
     if (!context.currentUserDetails) {
       throw redirect({ to: '/login' });
     }
   },
-  component: () => <CreateApartment />,
+  component: () => <SelectApartment />,
 });
