+++ apps/server/src/user/user.service.ts
@@ -57,6 +57,24 @@ export class UserService {
 
   async getUserByPhone(phoneNumber: string): Promise<User | null> {
     const user = await this.userRepo.findOneBy({ phoneNumber });
+
+    if (!user) {
+      this.logger.error(`User with phone number ${phoneNumber} not found`);
+      return null;
+    }
+    return user;
+  }
+
+  async getUserProfile(phoneNumber: string): Promise<User | null> {
+    const user = await this.userRepo.findOne({
+      where: { phoneNumber },
+      relations: {
+        apartments: {
+          apartment: true,
+        },
+      },
+    });
+
     if (!user) {
       this.logger.error(`User with phone number ${phoneNumber} not found`);
       return null;
