+++ libs/types/src/dto/user/create-user.dto.ts
@@ -1,3 +1,6 @@
+import { UserRole } from '../../enums';
+import { ApartmentInfoDto, ApartmentSettingsDto, RenterSettingsDto } from './create-apartment.dto';
+
 export class CreateUserDto {
   phoneNumber: string;
 
@@ -6,6 +9,20 @@ export class CreateUserDto {
   lastName: string;
 
   image?: string;
+}
+
+export type ApartmentDetails = { apartmentId: string } & ApartmentInfoDto & ApartmentSettingsDto & RenterSettingsDto;
+
+export interface Apartment {
+  apartment: ApartmentDetails;
+
+  createdAt: string;
+
+  rent: number;
+
+  role: UserRole;
+
+  userId: string;
 }
 
 export class UserResponse {
@@ -18,4 +35,6 @@ export class UserResponse {
   lastName: string;
 
   image?: string;
+
+  apartments: Apartment[];
 }
